Enable the Workspace to use analyzer options for Option APIs
This allows us to optionally use the Workspace itself as the source
of options intead of the Visual Studio service for .editorconfig. This
is still left under feature flag, so if the option is off the existing
library will be used. The CodeStyle layer and dotnet-format projects are
continuing to use the VS layer (for now).