Add an early asynchronous fetch of MEF before we try using it later
We have a bunch of places that try to fetch IComponentModel
synchronously as a part of our package initialization; this isn't good
since that loads the MEF cache which is potentially expensive and isn't
good to block on if we can avoid it. Since right now we can't free
our package init from using MEF, let's just fetch it asynchronously
at the very start.